pyproject.toml
[tool.poetry]
name = "hiddifypanel"
version = '10.80.0.dev9'
description = "hiddifypanel multi proxy panel"
readme = "README.md"
homepage = "https://hiddify.com"
repository = "https://github.com/hiddify/hiddify-manager/"
authors = ["hiddify" ]
license =  "LICENSE.md" 
keywords = ["proxy", "panel", "multi"]
documentation = "https://hiddify.com/manager"

[project.scripts]
hiddifypanel = "hiddifypanel.__main__:main"

[tool.setuptools.packages.find]
include = ["hiddifypanel"]
exclude = ["tests", ".github"]

[tool.setuptools.package-data]
"hiddifypanel" = [
    "translations/*/LC_MESSAGES/*.mo",
    "translations.i18n/*.json"
]
"" = ["hiddifypanel/translations/*/LC_MESSAGES/*.mo"]


[tool.poetry.dependencies]
python = "^3.10"
apiflask = {extras = ["async"], version = "^2.2.1"}
babel = "2.16.0"
bleach = "6.2.0"
bootstrap-flask = "2.4.1"
flask-admin = "1.6.1"
flask-adminlte3 = "1.0.9"
flask-classful = "0.16.0"
flask-login = "0.6.3"
flask-restful = "0.3.10"
flask-sqlalchemy = "3.1.1"
flask-session = "0.8.0"
flask-wtf = "1.2.2"
flask = "3.0.3"
markupsafe = "3.0.2"
pymysql = "1.1.1"
pyyaml = "6.0.2"
sqlalchemy-utils = "0.41.2"

sqlalchemy = {extras = ["asyncio"], version = "^2.0.29"}
strenum = "0.4.15"
wtforms = "3.1.2"
werkzeug = "3.0.6"
ansi2html = "1.9.2"
bjoern = "3.2.2"
click = "8.1.7"
cloudflare = "3.1.0"
cryptography = "<42"
fastenumplus = "1.4.0"
flask-apispec = "0.11.4"
flask-babel = "4.0.0"
loguru = "0.7.2"
maxminddb = "2.6.2"
netifaces = "0.11.0"
ping3 = "4.0.8"
psutil = "6.1.0"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
python-redis-cache = "4.0.1"
python-slugify = "8.0.4"
redis = "5.2.0"
requests = "2.32.3"
pytelegrambotapi = "4.23.0"
user-agents = "2.2.0"
xtlsapi = "3.2.1"
mysqlclient = "2.2.5"
cachecontrol = {version = "<0.15.0", extras = ["filecache"]}
sqlalchemy-serializer = "^1.4.22"
sonora = "^0.2.3"
protobuf = "^5.26.1"
asgiref = "^3.8.1"
celery = "^5.4.0"
dynaconf = "^3.2.6"


[tool.poetry.group.dev.dependencies]
pytest = "8.3.3"
coverage = "7.6.4"
flake8 = "7.1.1"
black = "24.10.0"
isort = "5.13.2"
pytest-cov = "6.0.0"
codecov = "2.1.13"
mypy = "1.13.0"
gitchangelog = "3.0.4"
mkdocs = "1.6.1"
flask-debugtoolbar = "0.16.0"
flask-shell-ipython = "0.5.3"
pytest-flask = "1.3.0"
grpcio-tools = "1.67.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
hiddifypanel = 'hiddifypanel.__main__:main'


[[tool.poetry.source]]
name = 'aliyun.mirrors' 
url = 'https://mirrors.aliyun.com/pypi/simple/' 
priority="supplemental"

[[tool.poetry.source]]
name = "pypi-public"
url = "https://pypi.org/simple/"
priority="primary"
